Summary characteristics of a panel revisit design

Panel revisit design characteristics are summarized: number of panels, number of time periods, total number of sample events for the revisit design, total number of sample events for each panel, total number of sample events for each time period and cumulative number of unique units sampled by time periods.

pd_summary(object, visitdsgn = NULL, ...)

Arguments

object

Two-dimensional array from panel_design and dimnames specifying revisit panel design. Typically, array is output from revisit_dsgn, revisit_bibd or revisit_rand functions.

visitdsgn

Two-dimensional array with same dimensions as paneldsgn specifying the number of times a sample unit is sampled at each time period. Default is visitdsgn=NULL, where default assumes that a sample unit will be sampled only once at each time period.

...

Additional arguments (S3 consistency)

Value

List of six elements.

n_panel

number of panels in revisit design

n_period

number of time periods in revisit design

n_total

total number of sample events across all panels and all time periods, accounting for visitdsgn, that will be sampled in the revisit design

n_periodunit

vector of the number of time periods a unit will be sampled in each panel

n_unitpnl

vector of the number of sample units, accounting for visitdsgn, that will be sampled in each panel

n_unitperiod

vector of the number of sample units, accounting for visitdsgn, that will be sampled during each time period

ncum_unit

vector of the cumulative number of unique units that will be sampled in time periods up to and including the current time period.

Details

The revisit panel design and the visit design (if present) are summarized. Summaries can be useful to know the effort required to complete the survey design. See the values returned for the summaries that are produced.

Author

Tony Olsen Olsen.Tony@epa.gov

Examples

# Serially alternating panel revisit design summary
sa_dsgn <- revisit_dsgn(20, panels = list(SA60N = list(
  n = 60, pnl_dsgn = c(1, 4),
  pnl_n = NA, start_option = "None"
)), begin = 1)
pd_summary(sa_dsgn)
#> $n_panel
#> [1] 5
#> 
#> $n_period
#> [1] 20
#> 
#> $n_total
#> [1] 1200
#> 
#> $n_periodunit
#> SA60N_1 SA60N_2 SA60N_3 SA60N_4 SA60N_5 
#>       4       4       4       4       4 
#> 
#> $n_unitpnl
#> SA60N_1 SA60N_2 SA60N_3 SA60N_4 SA60N_5 
#>     240     240     240     240     240 
#> 
#> $n_unitperiod
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#> 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 60 
#> 
#> $ncum_unit
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>  60 120 180 240 300 300 300 300 300 300 300 300 300 300 300 300 300 300 300 300 
#> 
# Add visit design where first panel is sampled twice at every time period
sa_visit <- sa_dsgn
sa_visit[sa_visit > 0] <- 1
sa_visit[1, sa_visit[1, ] > 0] <- 2
pd_summary(sa_dsgn, sa_visit)
#> $n_panel
#> [1] 5
#> 
#> $n_period
#> [1] 20
#> 
#> $n_total
#> [1] 1440
#> 
#> $n_periodunit
#> SA60N_1 SA60N_2 SA60N_3 SA60N_4 SA60N_5 
#>       8       4       4       4       4 
#> 
#> $n_unitpnl
#> SA60N_1 SA60N_2 SA60N_3 SA60N_4 SA60N_5 
#>     480     240     240     240     240 
#> 
#> $n_unitperiod
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#> 120  60  60  60  60 120  60  60  60  60 120  60  60  60  60 120  60  60  60  60 
#> 
#> $ncum_unit
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>  60 120 180 240 300 300 300 300 300 300 300 300 300 300 300 300 300 300 300 300 
#>