Create a balanced incomplete block panel revisit design

Create a revisit design for panels in a survey that specifies the time periods for the units of each panel to be sampled based on searching for a D-optimal block design that is a member of the class of generalized Youden designs. The resulting design need not be a balanced incomplete block design. Based on algorithmic idea by Cook and Nachtsheim (1989) and implemented by Robert Wheeler.

revisit_bibd(
  n_period,
  n_pnl,
  n_visit,
  nsamp,
  panel_name = "BIB",
  begin = 1,
  skip = 1,
  iter = 30
)

Arguments

n_period

Number of time periods for the survey design. Typically, number of periods if sampling occurs once per period or number of months if sampling occurs once per month. (v, number of varieties/treatments in BIBD terms)

n_pnl

Number of panels (b, number of blocks in BIBD terms)

n_visit

Number of time periods to be visited in a panel (k, block size in BIBD terms)

nsamp

Number of samples in each panel.

panel_name

Prefix for name of each panel

begin

Numeric name of first sampling occasion, e.g. a specific period.

skip

Number of sampling occasions to skip between planned sampling periods, e.g., sampling will occur only every 5 periods if skip = 5.

iter

Maximum number of iterations in search for D-optimal Generalized Youden Design.

Value

A two-dimensional array of sample sizes to be sampled for each panel and each sampling occasion.

Details

The function uses find.BIB function from crossdes package to search for a D-optimal block design. crossdes uses package AlgDesign to search balanced incomplete block designs.

References

Cook R. D. and C. Nachtsheim. (1989). Computer-aided blocking of factorial and response-surface designs. Technometrics 31(3), 339-346.

See also

Author

Tony Olsen Olsen.Tony@epa.gov

Examples

# Balanced incomplete block design with 20 sample occasions, 20 panels,
# 3 visits to each unit, and 20 units in each panel.
revisit_bibd(n_period = 20, n_pnl = 20, n_visit = 3, nsamp = 20)
#>         1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
#> BIB_01  0  0  0  0  0 20  0  0 20  0  0 20  0  0  0  0  0  0  0  0
#> BIB_02 20  0  0  0  0  0 20 20  0  0  0  0  0  0  0  0  0  0  0  0
#> BIB_03  0  0  0  0  0  0  0  0  0 20  0  0  0  0  0 20  0  0  0 20
#> BIB_04  0  0  0  0  0  0  0 20  0  0  0  0 20  0  0 20  0  0  0  0
#> BIB_05  0  0 20  0  0  0 20  0 20  0  0  0  0  0  0  0  0  0  0  0
#> BIB_06  0  0  0 20  0 20  0  0  0  0  0  0  0  0  0  0  0 20  0  0
#> BIB_07  0  0  0  0  0 20  0  0  0  0  0  0  0  0  0  0 20  0  0 20
#> BIB_08  0  0  0  0  0  0  0  0  0  0 20  0 20  0  0  0 20  0  0  0
#> BIB_09  0 20 20  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 20  0
#> BIB_10  0  0  0  0  0  0  0  0  0  0 20  0  0 20 20  0  0  0  0  0
#> BIB_11 20  0  0 20  0  0  0  0  0  0  0  0  0  0 20  0  0  0  0  0
#> BIB_12 20  0  0  0  0  0  0  0  0 20  0 20  0  0  0  0  0  0  0  0
#> BIB_13  0 20  0  0  0  0  0 20  0  0  0  0  0  0  0  0  0 20  0  0
#> BIB_14  0  0  0  0 20  0 20  0  0  0  0  0  0  0  0  0 20  0  0  0
#> BIB_15  0  0  0  0 20  0  0  0  0 20  0  0  0  0  0  0  0  0 20  0
#> BIB_16  0  0 20  0  0  0  0  0  0  0  0  0  0  0 20  0  0  0  0 20
#> BIB_17  0  0  0  0  0  0  0  0 20  0  0  0  0 20  0 20  0  0  0  0
#> BIB_18  0 20  0  0  0  0  0  0  0  0 20 20  0  0  0  0  0  0  0  0
#> BIB_19  0  0  0  0 20  0  0  0  0  0  0  0  0 20  0  0  0 20  0  0
#> BIB_20  0  0  0 20  0  0  0  0  0  0  0  0 20  0  0  0  0  0 20  0
#> attr(,"class")
#> [1] "paneldesign"