Print grts() and irs() errors.

This function prints the error messages vector in the grts and irs functions.

stopprnt(stop_df = get("stop_df", envir = .GlobalEnv), m = 1:nrow(stop_df))

Arguments

stop_df

Data frame that contains stop messages. The default is stop_df, which is the name given to the stop data frame created by functions in the spsurvey package.

m

Vector of indices for stop messages that are to be printed. The default is a vector containing the integers from 1 through the number of rows in stop_df, which will print all stop messages in the data frame.

Value

Printed errors

Author

Tony Olsen Olsen.Tony@epa.gov