Change analysis

This function organizes input and output for the estimation of change between two samples (for categorical and continuous variables). The analysis data, dframe, can be either a data frame or a simple features (sf) object. If an sf object is used, coordinates are extracted from the geometry column in the object, arguments xcoord and ycoord are assigned values "xcoord" and "ycoord", respectively, and the geometry column is dropped from the object.

change_analysis(
  dframe,
  vars_cat = NULL,
  vars_cont = NULL,
  test = "mean",
  subpops = NULL,
  surveyID = "surveyID",
  survey_names = NULL,
  siteID = "siteID",
  weight = "weight",
  revisitwgt = FALSE,
  xcoord = NULL,
  ycoord = NULL,
  stratumID = NULL,
  clusterID = NULL,
  weight1 = NULL,
  xcoord1 = NULL,
  ycoord1 = NULL,
  sizeweight = FALSE,
  sweight = NULL,
  sweight1 = NULL,
  fpc = NULL,
  popsize = NULL,
  vartype = "Local",
  jointprob = "overton",
  conf = 95,
  All_Sites = FALSE
)

Arguments

dframe

Data to be analyzed (analysis data). A data frame or sf object containing survey design variables, response variables, and subpopulation (domain) variables.

vars_cat

Vector composed of character values that identify the names of categorical response variables in dframe. The default is NULL.

vars_cont

Vector composed of character values that identify the names of continuous response variables in dframe. The default is NULL.

test

Character string or character vector providing the location measure(s) to use for change estimation for continuous variables. The choices are "mean", "median", or c("mean", "median"). The default is "mean".

subpops

Vector composed of character values that identify the names of subpopulation (domain) variables in dframe. If a value is not provided, the value "All_Sites" is assigned to the subpops argument and a factor variable named "All_Sites" that takes the value "All Sites" is added to dframe. The default value is NULL.

surveyID

Character value providing name of the survey ID variable in dframe. The default value is "surveyID".

survey_names

Character vector of length two that provides the survey names contained in the surveyID variable in the dframe data frame. The two values in the vector identify the first survey and second survey, respectively. If a value is not provided, unique values of the surveyID variable are assigned to the survey_names argument. The default is NULL.

siteID

Character value providing name of the site ID variable in dframe. For a two-stage sample, the site ID variable identifies stage two site IDs. The default value is "siteID".

weight

Character value providing name of the design weight variable in dframe. For a two-stage sample, the weight variable identifies stage two weights. The default value is "weight".

revisitwgt

Logical value that indicates whether each repeat visit site has the same design weight in the two surveys, where TRUE = the weight for each repeat visit site is the same and FALSE = the weight for each repeat visit site is not the same. When this argument is FALSE, all of the repeat visit sites are assigned equal weights when calculating the covariance component of the change estimate standard error. The default is FALSE.

xcoord

Character value providing name of the x-coordinate variable in dframe. For a two-stage sample, the x-coordinate variable identifies stage two x-coordinates. Note that x-coordinates are required for calculation of the local mean variance estimator. If dframe is an sf object, this argument is not required (as the geometry column in dframe is used to find the x-coordinate). The default value is NULL.

ycoord

Character value providing name of the y-coordinate variable in dframe. For a two-stage sample, the y-coordinate variable identifies stage two y-coordinates. Note that y-coordinates are required for calculation of the local mean variance estimator. If dframe is an sf object, this argument is not required (as the geometry column in dframe is used to find the y-coordinate). The default value is NULL.

stratumID

Character value providing name of the stratum ID variable in dframe. The default value is NULL.

clusterID

Character value providing the name of the cluster (stage one) ID variable in dframe. Note that cluster IDs are required for a two-stage sample. The default value is NULL.

weight1

Character value providing name of the stage one weight variable in dframe. The default value is NULL.

xcoord1

Character value providing the name of the stage one x-coordinate variable in dframe. Note that x coordinates are required for calculation of the local mean variance estimator. The default value is NULL.

ycoord1

Character value providing the name of the stage one y-coordinate variable in dframe. Note that y-coordinates are required for calculation of the local mean variance estimator. The default value is NULL.

sizeweight

Logical value that indicates whether size weights should be used during estimation, where TRUE uses size weights and FALSE does not use size weights. To employ size weights for a single-stage sample, a value must be supplied for argument weight. To employ size weights for a two-stage sample, values must be supplied for arguments weight and weight1. The default value is FALSE.

sweight

Character value providing the name of the size weight variable in dframe. For a two-stage sample, the size weight variable identifies stage two size weights. The default value is NULL.

sweight1

Character value providing name of the stage one size weight variable in dframe. The default value is NULL.

fpc

Object that specifies values required for calculation of the finite population correction factor used during variance estimation. The object must match the survey design in terms of stratification and whether the design is single-stage or two-stage. For an unstratified design, the object is a vector. The vector is composed of a single numeric value for a single-stage design. For a two-stage unstratified design, the object is a named vector containing one more than the number of clusters in the sample, where the first item in the vector specifies the number of clusters in the population and each subsequent item specifies the number of stage two units for the cluster. The name for the first item in the vector is arbitrary. Subsequent names in the vector identify clusters and must match the cluster IDs. For a stratified design, the object is a named list of vectors, where names must match the strata IDs. For each stratum, the format of the vector is identical to the format described for unstratified single-stage and two-stage designs. Note that the finite population correction factor is not used with the local mean variance estimator.

Example fpc for a single-stage unstratified survey design:

fpc <- 15000

Example fpc for a single-stage stratified survey design:

fpc <- list( Stratum_1 = 9000, Stratum_2 = 6000)

Example fpc for a two-stage unstratified survey design:

fpc <- c( Ncluster = 150, Cluster_1 = 150, Cluster_2 = 75, Cluster_3 = 75, Cluster_4 = 125, Cluster_5 = 75)

Example fpc for a two-stage stratified survey design:

fpc <- list( Stratum_1 = c( Ncluster_1 = 100, Cluster_1 = 125, Cluster_2 = 100, Cluster_3 = 100, Cluster_4 = 125, Cluster_5 = 50), Stratum_2 = c( Ncluster_2 = 50, Cluster_1 = 75, Cluster_2 = 150, Cluster_3 = 75, Cluster_4 = 75, Cluster_5 = 125))

popsize

Object that provides values for the population argument of the calibrate or postStratify functions in the survey package. If a value is provided for popsize, then either the calibrate or postStratify function is used to modify the survey design object that is required by functions in the survey package. Whether to use the calibrate or postStratify function is dictated by the format of popsize, which is discussed below. Post-stratification adjusts the sampling and replicate weights so that the joint distribution of a set of post-stratifying variables matches the known population joint distribution. Calibration, generalized raking, or GREG estimators generalize post-stratification and raking by calibrating a sample to the marginal totals of variables in a linear regression model. For the calibrate function, the object is a named list, where the names identify factor variables in dframe. Each element of the list is a named vector containing the population total for each level of the associated factor variable. For the postStratify function, the object is either a data frame, table, or xtabs object that provides the population total for all combinations of selected factor variables in the dframe data frame. If a data frame is used for popsize, the variable containing population totals must be the last variable in the data frame. If a table is used for popsize, the table must have named dimnames where the names identify factor variables in the dframe data frame. If the popsize argument is equal to NULL, then neither calibration nor post-stratification is performed. The default value is NULL.

Example popsize for calibration:

popsize <- list( Ecoregion = c( East = 750, Central = 500, West = 250), Type = c( Streams = 1150, Rivers = 350))

Example popsize for post-stratification using a data frame:

popsize <- data.frame( Ecoregion = rep(c("East", "Central", "West"), rep(2, 3)), Type = rep(c("Streams", "Rivers"), 3), Total = c(575, 175, 400, 100, 175, 75))

Example popsize for post-stratification using a table:

popsize <- with(MySurveyFrame, table(Ecoregion, Type))

Example popsize for post-stratification using an xtabs object:

popsize <- xtabs(~Ecoregion + Type, data = MySurveyFrame)

vartype

Character value providing the choice of the variance estimator, where "Local" indicates the local mean estimator and "SRS" indicates the simple random sampling estimator. The default value is "Local".

jointprob

Character value providing the choice of joint inclusion probability approximation for use with Horvitz-Thompson and Yates-Grundy variance estimators, where "overton" indicates the Overton approximation, "hr" indicates the Hartley-Rao approximation, and "brewer" equals the Brewer approximation. The default value is "overton".

conf

Numeric value providing the Gaussian-based confidence level. The default value is 95.

All_Sites

A logical variable used when subpops is not NULL. If All_Sites is TRUE, then alongside the subpopulation output, output for all sites (ignoring subpopulations) is returned for each variable in vars. If All_Sites is FALSE, then alongside the subpopulation output, output for all sites (ignoring subpopulations) is not returned for each variable in vars. The default is FALSE.

Value

List of change estimates composed of three items: (1) catsum contains change estimates for categorical variables, (2) contsum_mean contains estimates for continuous variables using the mean, and (3) contsum_median contains estimates for continuous variables using the median. The items in the list will contain NULL

for estimates that were not calculated. Each data frame includes estimates for all combinations of population Types, subpopulations within types, response variables, and categories within each response variable (for categorical variables and continuous variables using the median). Change estimates are provided plus standard error estimates and confidence interval estimates.

See also

Author

Tom Kincaid Kincaid.Tom@epa.gov

Examples

# Categorical variable example for three resource classes
dframe <- data.frame(
  surveyID = rep(c("Survey 1", "Survey 2"), c(100, 100)),
  siteID = paste0("Site", 1:200),
  wgt = runif(200, 10, 100),
  xcoord = runif(200),
  ycoord = runif(200),
  stratum = rep(rep(c("Stratum 1", "Stratum 2"), c(2, 2)), 50),
  CatVar = rep(c("North", "South"), 100),
  All_Sites = rep("All Sites", 200),
  Resource_Class = sample(c("Good", "Fair", "Poor"), 200, replace = TRUE)
)
myvars <- c("CatVar")
mysubpops <- c("All_Sites", "Resource_Class")
change_analysis(dframe,
  vars_cat = myvars, subpops = mysubpops,
  surveyID = "surveyID", siteID = "siteID", weight = "wgt",
  xcoord = "xcoord", ycoord = "ycoord", stratumID = "stratum"
)
#> $catsum
#>   Survey_1 Survey_2           Type Subpopulation Indicator Category  DiffEst.P
#> 1 Survey 1 Survey 2      All_Sites     All Sites    CatVar    North   3.000475
#> 2 Survey 1 Survey 2      All_Sites     All Sites    CatVar    South  -3.000475
#> 3 Survey 1 Survey 2 Resource_Class          Fair    CatVar    North  14.689313
#> 4 Survey 1 Survey 2 Resource_Class          Fair    CatVar    South -14.689313
#> 5 Survey 1 Survey 2 Resource_Class          Good    CatVar    North  -2.134971
#> 6 Survey 1 Survey 2 Resource_Class          Good    CatVar    South   2.134971
#> 7 Survey 1 Survey 2 Resource_Class          Poor    CatVar    North  -4.938677
#> 8 Survey 1 Survey 2 Resource_Class          Poor    CatVar    South   4.938677
#>   StdError.P MarginofError.P LCB95Pct.P UCB95Pct.P  DiffEst.U StdError.U
#> 1   6.585026        12.90641  -9.905938  15.906889  382.06004   387.8324
#> 2   6.585026        12.90641 -15.906889   9.905938   63.51599   385.1670
#> 3  11.769332        23.06747  -8.378153  37.756779  368.12859   233.8012
#> 4  11.769332        23.06747 -37.756779   8.378153 -151.88865   229.3587
#> 5  11.112683        21.78046 -23.915429  19.645488  249.50372   233.6940
#> 6  11.112683        21.78046 -19.645488  23.915429  308.38128   238.0019
#> 7  11.256382        22.06210 -27.000780  17.123425 -235.57227   196.7008
#> 8  11.256382        22.06210 -17.123425  27.000780  -92.97664   201.8435
#>   MarginofError.U LCB95Pct.U UCB95Pct.U nResp_1 Estimate.P_1 StdError.P_1
#> 1        760.1374 -378.07741  1142.1975      50     48.19580     4.611864
#> 2        754.9134 -691.39739   818.4294      50     51.80420     4.611864
#> 3        458.2420  -90.11337   826.3706      12     41.20733     8.243510
#> 4        449.5348 -601.42342   297.6461      19     58.79267     8.243510
#> 5        458.0317 -208.52802   707.5355      18     53.27583     7.996568
#> 6        466.4751 -158.09385   774.8564      16     46.72417     7.996568
#> 7        385.5264 -621.09867   149.9541      20     50.02829     7.193177
#> 8        395.6059 -488.58256   302.6293      15     49.97171     7.193177
#>   MarginofError.P_1 LCB95Pct.P_1 UCB95Pct.P_1 Estimate.U_1 StdError.U_1
#> 1          9.039088     39.15671     57.23488    2472.7231     249.6333
#> 2          9.039088     42.76512     60.84329    2657.8553     270.2843
#> 3         16.156982     25.05034     57.36431     693.6219     145.9089
#> 4         16.156982     42.63569     74.94966     989.6272     164.0714
#> 5         15.672986     37.60284     68.94881     893.4358     141.4375
#> 6         15.672986     31.05119     62.39716     783.5644     145.5566
#> 7         14.098368     35.92992     64.12666     885.6654     140.9146
#> 8         14.098368     35.87334     64.07008     884.6637     144.4921
#>   MarginofError.U_1 LCB95Pct.U_1 UCB95Pct.U_1 nResp_2 Estimate.P_2 StdError.P_2
#> 1          489.2722    1983.4509    2961.9953      50     51.19627     4.700348
#> 2          529.7475    2128.1078    3187.6028      50     48.80373     4.700348
#> 3          285.9761     407.6458     979.5981      17     55.89664     8.400102
#> 4          321.5741     668.0530    1311.2013      15     44.10336     8.400102
#> 5          277.2124     616.2233    1170.6482      20     51.14086     7.716646
#> 6          285.2858     498.2787    1068.8502      20     48.85914     7.716646
#> 7          276.1876     609.4778    1161.8530      13     45.08961     8.658194
#> 8          283.1993     601.4644    1167.8630      15     54.91039     8.658194
#>   MarginofError.P_2 LCB95Pct.P_2 UCB95Pct.P_2 Estimate.U_2 StdError.U_2
#> 1          9.212513     41.98376     60.40878    2854.7831     296.8117
#> 2          9.212513     39.59122     58.01624    2721.3713     274.4084
#> 3         16.463898     39.43274     72.36054    1061.7505     182.6845
#> 4         16.463898     27.63946     60.56726     837.7385     160.2684
#> 5         15.124348     36.01651     66.26521    1142.9395     186.0331
#> 6         15.124348     33.73479     63.98349    1091.9457     188.3034
#> 7         16.969749     28.11986     62.05936     650.0931     137.2379
#> 8         16.969749     37.94064     71.88014     791.6871     140.9355
#>   MarginofError.U_2 LCB95Pct.U_2 UCB95Pct.U_2
#> 1          581.7402    2273.0429    3436.5233
#> 2          537.8306    2183.5407    3259.2019
#> 3          358.0550     703.6956    1419.8055
#> 4          314.1204     523.6182    1151.8589
#> 5          364.6181     778.3214    1507.5576
#> 6          369.0678     722.8779    1461.0136
#> 7          268.9814     381.1117     919.0745
#> 8          276.2285     515.4585    1067.9156
#> 
#> $contsum_mean
#> NULL
#> 
#> $contsum_median
#> NULL
#>